tests/artifacts/slack-newsletter/requirements.txt
any-agent[all,a2a]==0.25.0
python-dotenv
beautifulsoup4
requests
fire